# Update Exists method to use HeadObjectWithContext