@marwan-at-work 2023-03-01
> OK. I think we can merge this for now. However, I feel we should change the directory layout to avoid putting too many files in one directory. Something like /github.com/aws/aws-sdk-go/@v/v1.44.90/v1.44.90.mod, /github.com/aws/aws-sdk-go/@v/v1.44.90/v1.44.90.info

Wouldn't this be a breaking change?