@linzhp 2023-03-01
It won't. The directory layout in a blob storage is implementation details of Athens and not visible to its clients. There is no API changes. The existing deployments will have a cache miss after the deploy the new layout. But the cache will be populated after that and no migration is needed.