@marwan-at-work 2023-03-01
> The existing deployments will have a cache miss after the deploy the new layout. But the cache will be populated after that and no migration is needed.

That's the breaking change no? It won't get the old (already stored) module, it will refetch it from the VCS causing two things: 

1. Duplication of a module version, we need to clean up and delete the old path now? (Or ideally rename it)
2. Immutability is broken: if the module has changed in the VCS, the checksums will fail.